vhost: add null callback checks
[ upstream commit 66be1a05a480a0484b02c1a0194126763f4dbb99 ]

rte_vhost_driver_callback_register() does not specify
any mandatory callbacks in struct rte_vhost_device_ops.

Add some missing NULL checks before calling rte_vhost_device_ops
callbacks.

Fixes: 4796ad63ba1f ("examples/vhost: import userspace vhost application")

+++ lib/vhost/vhost.c
@@ -750,10 +750,11 @@ vhost_destroy_device_notify(struct virtio_net *dev)
 
 	if (dev->flags & VIRTIO_DEV_RUNNING) {
 		vdpa_dev = dev->vdpa_dev;
-		if (vdpa_dev)
+		if (vdpa_dev && vdpa_dev->ops->dev_close)
 			vdpa_dev->ops->dev_close(dev->vid);
 		dev->flags &= ~VIRTIO_DEV_RUNNING;
-		dev->notify_ops->destroy_device(dev->vid);
+		if (dev->notify_ops->destroy_device)
+			dev->notify_ops->destroy_device(dev->vid);
 	}
 }
